commit flights

+++ safetrip-backend/src/main/java/com/safetripbackend/dto/CityResponseDto.java
@@ -1,12 +1,13 @@
 package com.safetripbackend.dto;
 
 import lombok.Data;
+import com.safetripbackend.dto.CityResponseDto;
 
 import java.util.List;
 
 @Data
 public class CityResponseDto {
-    private int id;
+    private Long id;
     private String name;
     private List<String> sights;
 }
